# Conflicts:
#	app/src/main/java/com/teamwss/websoso/data/mapper/UserMapper.kt
#	app/src/main/java/com/teamwss/websoso/data/remote/api/UserApi.kt
#	app/src/main/java/com/teamwss/websoso/data/repository/UserRepository.kt
#	app/src/main/java/com/teamwss/websoso/ui/main/myPage/MyPageFragment.kt
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/teamwss/websoso/ui/feedDetail/model/CommentModel.kt
@@ -1,11 +1,11 @@
 package com.teamwss.websoso.ui.feedDetail.model
 
-import com.teamwss.websoso.ui.feed.model.FeedModel.UserModel
+import com.teamwss.websoso.ui.main.feed.model.FeedModel.UserModel
 
 data class CommentModel(
     val user: UserModel,
     val commentContent: String,
-    val commentId: Int,
+    val commentId: Long,
     val createdDate: String,
     val isModified: Boolean,
     val isMyComment: Boolean,
